Plugin editor end programming
