+++ Source/PluginProcessor.cpp
@@ -363,3 +363,4 @@ juce::AudioProcessor* JUCE_CALLTYPE createPluginFilter()
 {
     return new SimpleEQAudioProcessor();
 }
+
